Первое домашнее задание
